app/requirements.txt
alembic==0.8.9
bcrypt==3.1.2
celery==4.0.2
cffi==1.9.1
click==6.6
coverage==4.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Dance==0.14.0
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
Flask==0.12
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.6
MarkupSafe==0.23
mongoengine==0.15.0
numpy==1.14.1
psycopg2==2.6.2
pycparser==2.17
PyJWT==1.4.2
pymongo==3.4.0
python-dotenv==0.7.1
python-editor==1.0.3
redis==2.10.5
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
SQLAlchemy==1.1.4
tweepy==3.6.0
Werkzeug==0.11.13